Removed the generic that was used in the native APDUTransport struct by replacing it with a trait object in order to be still "generic" regardless the type that implement the Exchange trait
Updated some functions that aimed to be generic over the APDUTransport

Install protoc before checks cause it is a dependency for ledger-transport-zemu

Fixes an issue with cargo fmt which tried to check modules that are  generated by protoc-rust.

Moved the build and build --release commands up so they are executed before formatting and clippy.
by doing so we can check compilation errors which might be hided if errors like formatting are triggered first.

Reduces the number of cfg tags by adding an inner module where the Exchange trait is declare and implemented on transport types